+++ src/containers/App.tsx
@@ -5,7 +5,7 @@ import { connect } from "react-redux";
 import { bootstrap } from "../actions/app";
 import { IRootReducerState } from "../reducers";
 import TransactionList from "../components/TransactionList";
-import PrivateRoute from "../components/PrivateRoute";
+import PrivateRoute from "./PrivateRoute";
 import MainContainer from "../containers/MainContainer";
 import SignIn from "../containers/SignIn";
 import SignUp from "../containers/SignUp";
